Formatting dates in miliseconds
